cleaned up registry.py, WIP. Also renamed it to registry_docker, will make a new registry file for local run
